requirements.txt
python-telegram-bot==20.0
whois==0.9.14
